+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Development]
 
+### Added
+* GFQL predicate `is_year_end`
+
 ### Docs
 
 * GFQL in readme.md