+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Development]
 
+## [0.32.0 - 2023-12-22]
+
 ### Added
 
 * GFQL `Chain` AST object